Added new RunRule to MetaAnimation.  You may now run all the animations in a MetaAnimation in a SEQUENCE or SIMULTANEOUSly.  Added a FinishRule option to MoveAnimation.  You may now run have a MoveAnimation terminate when only the X or Y are FIRST met, or if BOTH of them are.
Modified the Group class so that you can specify a show and hide animation.  Used them in the menu (which can be seen by changing some minor code).  Extracted the Interface from the Group class and made an abstract skeleton of it.

Modified ImmutableRectangle so that it has an informative toString method.  Added a way to disable the LayerManager and Entities in general.