Added item collision detection!! (finally).
Currently it will print to the console all the collisions. The collisions are simulated in advance and it calculates all possible scenarios achieved. This means that if a bomb hits a rocket and multiplier simultaneously and the rocket is pointed at the multiplier, the achievment would be thus:  bomb->x2->end (bomb into mult),  bomb->rocket->x2->end (bomb into rocket into mult) and finally bomb->Rocket-End (bomb into rocket).  This would achieve any of these 3 possible achievments. (this seems a bit weird because graphically it shows that the bomb detonated both the mult and the rocket simultaneously and the rocket did not actually hit the mult since the mult was gone at the time.. i think this is ok. what do you think cam?